xcode-software/src/main/java/com/xcode_software/controller/dto/ApiResponse.java
package com.xcode_software.controller.dto;

public class ApiResponse<T> {
    private final boolean success;
    private final String message;
    private final int statusCode;
    private final T details;

    private ApiResponse(Builder<T> builder) {
        this.success = builder.success;
        this.message = builder.message;
        this.statusCode = builder.statusCode;
        this.details = builder.details;
    }

    public boolean isSuccess() {
        return success;
    }

    public String getMessage() {
        return message;
    }

    public int getStatusCode() {
        return statusCode;
    }

    public T getDetails() {
        return details;
    }

    public static <T> Builder<T> builder() {
        return new Builder<>();
    }

    public static <T> ApiResponse<T> success(String message, T details) {
        return ApiResponse.<T>builder()
                .success(true)
                .statusCode(200)
                .message(message)
                .details(details)
                .build();
    }

    public static <T> ApiResponse<T> failure(String message, int statusCode) {
        return ApiResponse.<T>builder()
                .success(false)
                .statusCode(statusCode)
                .message(message)
                .build();
    }

    public static class Builder<T> {
        private boolean success;
        private String message;
        private int statusCode;
        private T details;

        public Builder<T> success(boolean success) {
            this.success = success;
            return this;
        }

        public Builder<T> message(String message) {
            this.message = message;
            return this;
        }

        public Builder<T> statusCode(int statusCode) {
            this.statusCode = statusCode;
            return this;
        }

        public Builder<T> details(T details) {
            this.details = details;
            return this;
        }

        public ApiResponse<T> build() {
            return new ApiResponse<>(this);
        }
    }
}
